Fix compile errors with VS2019 v16.9.3 and GCC v7.5.0 (see issue #3140)

+++ libcef_dll/base/cef_logging.cc
@@ -210,7 +210,7 @@ std::string SystemErrorCodeToString(SystemErrorCode error_code) {
   char msgbuf[error_message_buffer_size];
   DWORD flags = FORMAT_MESSAGE_FROM_SYSTEM | FORMAT_MESSAGE_IGNORE_INSERTS;
   DWORD len = FormatMessageA(flags, NULL, error_code, 0, msgbuf,
-                             base::size(msgbuf), NULL);
+                             static_cast<DWORD>(base::size(msgbuf)), NULL);
   std::stringstream ss;
   if (len) {
     std::string s(msgbuf);
